feat: [sc-9794] Add Arbitrum Support in FE app (#468)
* v0.5.1-arbitrum.1

* Add AAVE_V3 support for ARBITRUM network

The commit adds support for the ARBITRUM network in get-abi-for-contract.ts within the dma-library. Specific AAVE_V3 contract data (PoolDataProvider, LendingPool, and Oracle) have been included, to accommodate new functionality and facilitate future interactions with the ARBITRUM network.

* Add AAVE_V3 support for ARBITRUM network

The commit adds support for the ARBITRUM network in get-abi-for-contract.ts within the dma-library. Specific AAVE_V3 contract data (PoolDataProvider, LendingPool, and Oracle) have been included, to accommodate new functionality and facilitate future interactions with the ARBITRUM network.

* v0.1.1-arbitrum.1

* Adjust flashloan provider logic for 'Spark' protocol

This commit adjusts the conditional check in the build-operation.ts file for the 'flashloanProvider'. The previous code considered only 'Balancer' as the flashloanProvider. This has been expanded to include cases where the protocolType is 'Spark', enhancing the strategic options within the AAVE-like multiplication strategy for the ARBITRUM network.

* v0.5.1-arbitrum.2

* Fix flashloan condition for Spark protocol in DMA library

This commit modifies the condition for flashloanProvider in the AAVE-like multiplication strategy. Previously, the condition only checked if the provider was 'Balancer'. Now, it also checks if the protocolType is 'Spark'. This change allows more strategic options for flashloans in the ARBITRUM network when utilizing the 'Spark' protocol.

* v0.5.1-arbitrum.3