Missing tokens for Arbitrum  (#474)
* Correct addresses for DAI, RETH, WBTC in arbitrum config

Updated the address for DAI, RETH, and WBTC tokens in the arbitrum configuration file. The previous addresses were incorrect. Ensuring the correct addresses are set is critical, as it affects operations such as token transfers, oracles, and liquidity provision, which rely on accurate token contract addresses.

* v0.1.4-rc-arbitrum.1

* Update package version in addresses

The package version of the addresses component has been updated from "0.1.4-rc-arbitrum.1" to "0.1.4".