Add flashloan data to strategy
Introduced a new data structure, WithFlashloan, in types/strategies.ts. This structure is to hold flashloan data. The WithFlashloan data type is now used in where the IStrategy simulation object is used. Refreshed other packages to match new structure and types. This change is made in order to have a clear and separated identification and handling of flashloan data across the entire application procedure.