Update withdrawal amount in AAVE v3 close operation
Modified the withdrawal amount in AAVE v3 close operation to emit the given collateral amount instead of the previous maximum uint value, which may cause issues for larger withdrawals. This is essential to accurately relay the amount that will be withdrawn and to ensure precise operation of the AAVE strategy.