feat: [sc-14954] Use Flashmint module on spark if debt token is dai (#649)
* Implement conditioning for flashloan provider selection

The commit introduces an additional debtToken parameter in the resolveFlashloanProvider function to facilitate better decision making when selecting a flashloan provider. Further, the implementation specifies Balancer as the FlashloanProvider when the lendingProtocol is 'Ajna', or when it's 'Spark' with a debtToken not equivalent to 'DAI'. This change refines the flashloan provider selection process, enhancing the overall strategy execution.

* v0.6.39-rc.4

* v0.6.40

* Replace hard-coded FlashloanProvider with dynamic selection

The commit substitutes the hard-coded Balancer FlashloanProvider with the dynamic "flashloanProvider" variable within the Aave-like strategies. The objective is enhancing code maintainability and readability, also ensuring flexibility in situations where the protocolType is 'Spark'.